src/python/pyproject.toml
[tool.poetry]
name = "keys python reference"
version = "0.1.0"
description = "Keys python references and wikis"
authors = ["shangar-arivazhagan <shangar.arivazhagan@solitontech.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
grpcio = "^1.64.1"
grpcio-tools = "^1.64.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.8.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
